requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.0
astroid==2.15.5
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
dill==0.3.6
Django==4.2.1
django-cors-headers==4.0.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
frozenlist==1.3.3
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
multidict==6.0.4
openai==0.27.7
Pillow==9.5.0
platformdirs==3.5.1
pycodestyle==2.10.0
PyJWT==2.7.0
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
tqdm==4.65.0
typing_extensions==4.6.1
tzdata==2023.3
urllib3==2.0.2
wrapt==1.15.0
yarl==1.9.2
